Refactors Bot Modules
